Fixed more warnings on MSVC

+++ src/engine/core/src/runtime/md5digest.cc
@@ -12,13 +12,13 @@ extern"C"
 namespace BugEngine
 {
 
-MD5 digest(const void* buffer, size_t size)
+MD5 digest(const void* buffer, u64 size)
 {
     MD5_CTX context;
     MD5Init(&context);
     MD5Update(&context, (unsigned char*)buffer, be_checked_numcast<unsigned int>(size));
     MD5Final(&context);
-	MD5 result = { {0, 0, 0, 0} };
+    MD5 result = { {0, 0, 0, 0} };
 
     for (int i = 0; i < 4; ++i)
         result.hash[i] = context.digest[i*4+0] << 0
